Fix v7.1.2 Incident Command Wire live rotation
