+++ .github/fixtures/main-style-source-headroom.json
@@ -82,14 +82,14 @@
   ],
   "v7Candidate": {
     "issue": 519,
-    "version": "7.1.1",
-    "sourceBytes": 1554857,
-    "sourceLines": 23430,
-    "sourceSha256": "d02e064a0fa3a6bed48c4f08e226fa8d6e37c1411ea407d325150ea30bd7acc5",
-    "templateBytes": 644561,
+    "version": "7.1.2",
+    "sourceBytes": 1555226,
+    "sourceLines": 23433,
+    "sourceSha256": "47d7b384da48957f5c6a7615ed6a6e0aa633962161c64159890f1381f1294765",
+    "templateBytes": 644944,
     "templateLines": 7279,
-    "templateSha256": "3ef90a1a294c0fb085f169149913bd74370307e87ecef07c889f3da1a444b18d",
-    "canonicalCssSha256": "286d48620da3838940f7d2e9abd589177d69a5b2bf8baaa8dc02e2d678853adf",
+    "templateSha256": "946d1e9875a570734dfbc20caeaa8e50b895b326c791495dfb97668fceae9526",
+    "canonicalCssSha256": "d48f9698b54e8f0f7e92d4e521a9e735ae50872b8412a53165cb3561080103fc",
     "maxSourceBytes": 1600000,
     "minTemplateBytes": 500000,
     "retiredSystems": [
