--- .github/workflows/apply-issue519-v712-hotfix.yml
@@ -1,98 +1,0 @@
-name: Apply Issue 519 v7.1.2 Incident Wire Hotfix
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/issue519/apply_v712_hotfix.py"
-      - ".github/workflows/apply-issue519-v712-hotfix.yml"
-
-permissions:
-  contents: read
-
-concurrency:
-  group: toolkit-issue-519-v712-hotfix
-  cancel-in-progress: false
-
-jobs:
-  apply:
-    if: github.actor == 'Conroy1988' && github.head_ref == 'fix/issue-519-v712-dedicated-rotation'
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    steps:
-      - name: Check out exact v7.1.2 branch
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          ref: fix/issue-519-v712-dedicated-rotation
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Apply dedicated rotation and control containment repair
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 .github/issue519/apply_v712_hotfix.py
-          python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/scripts/test_issue517_incident_command_wire.py')
-          text = path.read_text(encoding='utf-8')
-          old = "    assert 'majorIncidentFeedScheduleAdvance(feed, 650);' in pause\n"
-          new = "    assert 'majorIncidentFeedScheduleAdvance(feed, 650, true);' in pause\n"
-          if text.count(old) != 1:
-              raise SystemExit(f'Expected one stale resume assertion, found {text.count(old)}')
-          path.write_text(text.replace(old, new, 1), encoding='utf-8')
-          PY
-
-      - name: Validate complete v7.1.2 candidate
-        shell: bash
-        run: |
-          set -euo pipefail
-          {
-            python3 .github/scripts/validate_userscript.py &&
-            bash .github/scripts/run_userscript_preflight.sh --contracts &&
-            python3 .github/scripts/test_documentation_consistency.py &&
-            node --check src/MissionChief_Map_Command_Toolkit.user.js &&
-            cmp --silent dist/MissionChief_Map_Command_Toolkit.user.js dist/MissionChief_Map_Command_Toolkit.txt
-          } 2>&1 | tee issue519-v712-validation.log
-
-      - name: Upload v7.1.2 validation diagnostic
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: issue519-v712-validation-${{ github.sha }}
-          path: issue519-v712-validation.log
-          if-no-files-found: error
-          retention-days: 14
-
-      - name: Enforce unchanged performance envelope
-        shell: bash
-        run: |
-          set -euo pipefail
-          git fetch origin main --depth=1
-          git show origin/main:src/MissionChief_Map_Command_Toolkit.user.js > /tmp/performance-base.user.js
-          python3 .github/scripts/test_performance_budget.py
-          python3 .github/scripts/check_performance_budget.py \
-            --candidate src/MissionChief_Map_Command_Toolkit.user.js \
-            --base /tmp/performance-base.user.js \
-            --policy .github/performance-budget.json \
-            --json-output /tmp/issue519-v712-performance.json \
-            --markdown-output /tmp/issue519-v712-performance.md
-          cat /tmp/issue519-v712-performance.md
-
-      - name: Commit exact v7.1.2 candidate
-        env:
-          DEVELOPMENT_PR_TOKEN: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          test -n "$DEVELOPMENT_PR_TOKEN"
-          test "$GITHUB_HEAD_REF" = "fix/issue-519-v712-dedicated-rotation"
-          git config user.name "Conroy1988"
-          git config user.email "27301455+Conroy1988@users.noreply.github.com"
-          rm -f issue519-v712-validation.log
-          git add -A
-          git diff --cached --quiet && { echo "No v7.1.2 hotfix changes produced."; exit 1; }
-          git commit -m "Fix v7.1.2 Incident Command Wire live rotation"
-          BASIC_AUTH="$(printf 'x-access-token:%s' "$DEVELOPMENT_PR_TOKEN" | base64 -w0)"
-          git -c http.https://github.com/.extraheader="AUTHORIZATION: basic ${BASIC_AUTH}" \
-            push origin HEAD:refs/heads/fix/issue-519-v712-dedicated-rotation
